refactor file name to not contain underscore
